Move qt5web classes to qt6webdisplay
They were used by both qt5 and qt6, now only for qt6
