+++ gui/CMakeLists.txt
@@ -18,9 +18,6 @@ if(webgui)
    if(cefweb)
       add_subdirectory(cefdisplay)
    endif()
-   if(qt5web)
-      add_subdirectory(qt5webdisplay)
-   endif()
    if(qt6web)
       add_subdirectory(qt6webdisplay)
    endif()
